requirements.txt
unstructured[all-docs]
pillow
lxml
chromadb
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
nltk
pytest
black
isort
IPython